requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3
nltk==3.8.1
spacy==3.7.2
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3
python-dotenv==1.0.0
whitenoise==6.6.0
